Removed generic in place of union
