+++ cvtModel/src/cvt_simulator/models/dataTypes.py
@@ -1,10 +1,6 @@
 from dataclasses import dataclass
 from typing import Union
 
-# Import the new PulleyState from the interface
-from cvt_simulator.models.pulley.pulley_interface import PulleyState
-
-
 ## Pulley stuff
 @dataclass
 class flyweightForceBreakdown:
@@ -53,18 +49,43 @@ class SecondaryForceBreakdown:
     helix_force: HelixForceBreakdown
     net: float
 
+# All possible pulley breakdown types
+PulleyBreakdowns = Union[PrimaryForceBreakdown, SecondaryForceBreakdown]
 
 @dataclass
-class RadialPulleyForceBreakdown:
-    pulleyForce: Union[PrimaryForceBreakdown, SecondaryForceBreakdown]
-    wrap_angle: float
-    radius: float
-    angular_velocity: float
-    radial_from_clamping: float
-    radial_from_centrifugal: float
+class PulleyForces:
+    """
+    Core outputs that every pulley must provide.
+    
+    These three values are sufficient to drive the CVT simulation regardless
+    of the internal mechanism (flyweights, helix, PID, etc.).
+    """
+    clamping_force: float  # Axial force pushing pulley halves together [N]
+    radial_force: float    # Total radial force on belt [N]
+    max_torque: float      # Maximum transferable torque before slip [N⋅m]
 
-    net: float
 
+@dataclass
+class PulleyState:
+    """
+    Complete pulley state including forces and geometric properties.
+    
+    Combines the core forces with geometric data needed for system-level
+    calculations (slip model, shift dynamics, etc.).
+    """
+    forces: PulleyForces
+    
+    # Geometric properties at current shift position
+    wrap_angle: float           # Belt wrap angle around pulley [rad]
+    radius: float              # Effective pitch radius [m]
+    angular_velocity: float    # Pulley angular velocity [rad/s]
+    
+    # Force components (for analysis/debugging)
+    radial_from_clamping: float      # Radial force contribution from clamping [N]
+    radial_from_centrifugal: float   # Radial force from belt centrifugal effect [N]
+    
+    # Implementation-specific breakdown (Union of all concrete breakdown types)
+    breakdown: PulleyBreakdowns
 
 @dataclass
 class CvtSystemForceBreakdown:
