+++ cvtModel/src/cvt_simulator/models/pulley/secondary_pulley_interface.py
@@ -1,6 +1,6 @@
 
 
-from cvt_simulator.models.pulley.pulley_interface import PulleyModel, TBreakdown
+from cvt_simulator.models.pulley.pulley_interface import PulleyModel
 from pyparsing import ABC
 from cvt_simulator.constants.car_specs import (
     BELT_HEIGHT,
@@ -11,7 +11,7 @@ from cvt_simulator.utils.system_state import SystemState
 from cvt_simulator.utils.theoretical_models import TheoreticalModels as tm
 
 
-class SecondaryPulleyModel(PulleyModel[TBreakdown], ABC):
+class SecondaryPulleyModel(PulleyModel, ABC):
     """
     Abstract base for secondary (driven-side) pulley implementations.
